split out common extract functions
