+++ src/ExportProvider/IBKR/Extract.py
@@ -1,27 +1,10 @@
 from lxml import etree
 import src.ExportProvider.IBKR.Schemas as s
-import arrow
+import src.ExportProvider.Utils as utils
 from typing import Any
 
 
-def safeDateParse(dateString: str) -> arrow.Arrow:
-     safeDateString = dateString.replace('EDT', 'UTC-4').replace('EST', 'UTC-5')
-     return arrow.get(safeDateString, ['YYYY-MM-DD HH:mm:ss ZZZ', 'YYYY-MM-DD;HH:mm:ss ZZZ', 'YYYY-MM-DD'])
 
-def valueOrNone(value: str) -> str | None:
-    if value == "":
-         return None
-    return value
-
-def floatValueOrNone(value: str) -> float | None:
-    if value == "":
-         return None
-    return float(value)
-
-def dateValueOrNone(value: str) -> arrow.Arrow | None:
-    if value == "":
-         return None
-    return safeDateParse(value)
 
 def deduplicateList(lines: list[list[Any]]):
     allRows = [x for xs in lines for x in xs]
@@ -51,13 +34,13 @@ def extractStockTrade(node: etree.ElementBase) -> s.TradeStock:
         Conid = node.attrib['conid'],
         SecurityID = node.attrib['securityID'],
         SecurityIDType = s.SecurityIDType(node.attrib['securityIDType']),
-        CUSIP = valueOrNone(node.attrib['cusip']),
+        CUSIP = utils.valueOrNone(node.attrib['cusip']),
         ISIN = node.attrib['isin'],
         FIGI = node.attrib['figi'],
         ListingExchange = node.attrib['listingExchange'],
-        ReportDate = safeDateParse(node.attrib['reportDate']),
-        DateTime = safeDateParse(node.attrib['dateTime']),
-        TradeDate = safeDateParse(node.attrib['tradeDate']),
+        ReportDate = utils.safeDateParse(node.attrib['reportDate']),
+        DateTime = utils.safeDateParse(node.attrib['dateTime']),
+        TradeDate = utils.safeDateParse(node.attrib['tradeDate']),
         TransactionType = s.TransactionType(node.attrib['transactionType']),
         Exchange = node.attrib['exchange'],
         Quantity = float(node.attrib['quantity']),
@@ -79,7 +62,7 @@ def extractStockTrade(node: etree.ElementBase) -> s.TradeStock:
         MarketToMarketProfitAndLoss = float(node.attrib['mtmPnl']),
         BuyOrSell = s.BuyOrSell(node.attrib['buySell']),
         TransactionID = node.attrib['transactionID'],
-        OrderTime = safeDateParse(node.attrib['orderTime']),
+        OrderTime = utils.safeDateParse(node.attrib['orderTime']),
         LevelOfDetail = s.LevelOfDetail(node.attrib['levelOfDetail']),
         ChangeInPrice = float(node.attrib['changeInPrice']),
         ChangeInQuantity = float(node.attrib['changeInQuantity']),
@@ -104,14 +87,14 @@ def extractStockLot(node: etree.ElementBase) -> s.LotStock:
         Conid = node.attrib['conid'],
         SecurityID = node.attrib['securityID'],
         SecurityIDType = s.SecurityIDType(node.attrib['securityIDType']),
-        CUSIP = valueOrNone(node.attrib['cusip']),
+        CUSIP = utils.valueOrNone(node.attrib['cusip']),
         ISIN = node.attrib['isin'],
         FIGI = node.attrib['figi'],
         ListingExchange = node.attrib['listingExchange'],
         Multiplier = float(node.attrib['multiplier']),
-        ReportDate = safeDateParse(node.attrib['reportDate']),
-        DateTime = safeDateParse(node.attrib['dateTime']),
-        TradeDate = safeDateParse(node.attrib['tradeDate']),
+        ReportDate = utils.safeDateParse(node.attrib['reportDate']),
+        DateTime = utils.safeDateParse(node.attrib['dateTime']),
+        TradeDate = utils.safeDateParse(node.attrib['tradeDate']),
         Exchange = node.attrib['exchange'],
         Quantity = float(node.attrib['quantity']),
         TradePrice = float(node.attrib['tradePrice']),
@@ -123,8 +106,8 @@ def extractStockLot(node: etree.ElementBase) -> s.LotStock:
         ForexProfitAndLoss = float(node.attrib['fxPnl']),
         BuyOrSell = s.BuyOrSell(node.attrib['buySell']),
         TransactionID = node.attrib['transactionID'],
-        OpenDateTime = safeDateParse(node.attrib['openDateTime']),
-        HoldingPeriodDateTime = safeDateParse(node.attrib['holdingPeriodDateTime']),
+        OpenDateTime = utils.safeDateParse(node.attrib['openDateTime']),
+        HoldingPeriodDateTime = utils.safeDateParse(node.attrib['holdingPeriodDateTime']),
         LevelOfDetail = s.LevelOfDetail(node.attrib['levelOfDetail']),
     )
     return trade
@@ -151,12 +134,12 @@ def extractOptionTrade(node: etree.ElementBase) -> s.TradeDerivative:
         TradeID = node.attrib['tradeID'],
         Multiplier = float(node.attrib['multiplier']),
         Strike = float(node.attrib['strike']),
-        ReportDate = safeDateParse(node.attrib['reportDate']),
-        Expiry = safeDateParse(node.attrib['expiry']),
-        DateTime = safeDateParse(node.attrib['dateTime']),
+        ReportDate = utils.safeDateParse(node.attrib['reportDate']),
+        Expiry = utils.safeDateParse(node.attrib['expiry']),
+        DateTime = utils.safeDateParse(node.attrib['dateTime']),
         PutOrCall = s.PutOrCall(node.attrib['putCall']),
-        TradeDate = safeDateParse(node.attrib['tradeDate']),
-        SettleDateTarget = safeDateParse(node.attrib['settleDateTarget']),
+        TradeDate = utils.safeDateParse(node.attrib['tradeDate']),
+        SettleDateTarget = utils.safeDateParse(node.attrib['settleDateTarget']),
         TransactionType = s.TransactionType(node.attrib['transactionType']),
         Exchange = node.attrib['exchange'],
         Quantity = float(node.attrib['quantity']),
@@ -178,7 +161,7 @@ def extractOptionTrade(node: etree.ElementBase) -> s.TradeDerivative:
         MarketToMarketProfitAndLoss = float(node.attrib['mtmPnl']),
         BuyOrSell = s.BuyOrSell(node.attrib['buySell']),
         TransactionID = node.attrib['transactionID'],
-        OrderTime = safeDateParse(node.attrib['orderTime']),
+        OrderTime = utils.safeDateParse(node.attrib['orderTime']),
         LevelOfDetail = s.LevelOfDetail(node.attrib['levelOfDetail']),
         OrderType = s.OrderType(node.attrib['orderType']),
     )
@@ -206,11 +189,11 @@ def extractOptionLot(node: etree.ElementBase) -> s.LotDerivative:
         UnderlyingListingExchange = node.attrib['underlyingListingExchange'],
         Multiplier = float(node.attrib['multiplier']),
         Strike = float(node.attrib['strike']),
-        ReportDate = safeDateParse(node.attrib['reportDate']),
-        Expiry = safeDateParse(node.attrib['expiry']),
-        DateTime = safeDateParse(node.attrib['dateTime']),
+        ReportDate = utils.safeDateParse(node.attrib['reportDate']),
+        Expiry = utils.safeDateParse(node.attrib['expiry']),
+        DateTime = utils.safeDateParse(node.attrib['dateTime']),
         PutOrCall = s.PutOrCall(node.attrib['putCall']),
-        TradeDate = safeDateParse(node.attrib['tradeDate']),
+        TradeDate = utils.safeDateParse(node.attrib['tradeDate']),
         Exchange = node.attrib['exchange'],
         Quantity = float(node.attrib['quantity']),
         TradePrice = float(node.attrib['tradePrice']),
@@ -222,8 +205,8 @@ def extractOptionLot(node: etree.ElementBase) -> s.LotDerivative:
         ForexProfitAndLoss = float(node.attrib['fxPnl']),
         BuyOrSell = s.BuyOrSell(node.attrib['buySell']),
         TransactionID = node.attrib['transactionID'],
-        OpenDateTime = safeDateParse(node.attrib['openDateTime']),
-        HoldingPeriodDateTime = safeDateParse(node.attrib['holdingPeriodDateTime']),
+        OpenDateTime = utils.safeDateParse(node.attrib['openDateTime']),
+        HoldingPeriodDateTime = utils.safeDateParse(node.attrib['holdingPeriodDateTime']),
         LevelOfDetail = s.LevelOfDetail(node.attrib['levelOfDetail']),
     )
     return trade
@@ -246,13 +229,13 @@ def extractCashTransaction(node: etree.ElementBase) -> s.TransactionCash:
         ISIN=node.attrib["isin"],
         FIGI=node.attrib["figi"],
         ListingExchange=node.attrib["listingExchange"],
-        DateTime=safeDateParse(str(node.attrib["dateTime"])),
-        SettleDate=safeDateParse(str(node.attrib["settleDate"])),
+        DateTime=utils.safeDateParse(str(node.attrib["dateTime"])),
+        SettleDate=utils.safeDateParse(str(node.attrib["settleDate"])),
         Amount=float(node.attrib["amount"]),
         Type=s.CashTransactionType(node.attrib["type"]),
         Code=node.attrib["code"],
         TransactionID=node.attrib["transactionID"],
-        ReportDate=safeDateParse(str(node.attrib["reportDate"])),
+        ReportDate=utils.safeDateParse(str(node.attrib["reportDate"])),
         ActionID=node.attrib["actionID"]
     )
     return trade
